refactor: move `Probot` code to `src/probot.ts`
